feat: MyPageScreen과 FooriendScreen 구현
ProfileSection과 ReviewLazyGrid를 Component로 뺐습니다.
usage
ProfileSection - MyPageScreen, FooriendScreen
ReviewLazyGrid - SocialScreen, MyPageScreen, FooriendScreen
MyPageScreen과 Fooriend Screen가 더 깔끔해지도록 구현했습니다.
