+++ frontend/app/src/main/java/com/team13/fooriend/ui/screen/FooriendScreen.kt
@@ -10,7 +10,10 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.team13.fooriend.ui.component.ProfileSection
+import com.team13.fooriend.ui.component.ReviewLazyGrid
 
 @Composable
 fun FooriendScreen(
@@ -23,10 +26,18 @@ fun FooriendScreen(
         IconButton(onClick = onBackClick) {
             Icon(Icons.Default.ArrowBack, contentDescription = "back")
         }
-        Text("FooriendScreen")
-        Button(onClick = onFollowClick){
-            Text("Follow")
-        }
+        
+        // profile section
+        ProfileSection(onFollowClick = onFollowClick)
+        
+        // review lazy grid
+        ReviewLazyGrid(onReviewClick = { })
     }
+}
+
+@Composable
+@Preview(showSystemUi = true, showBackground = true)
+fun FooriendScreenPreview(){
+    FooriendScreen()
 }
